backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.29
SQLAlchemy>=2.0
psycopg[binary]>=3.1
pydantic>=2.5
pydantic-settings>=2.2
httpx>=0.26
faker>=24.0
python-jose[cryptography]>=3.3
python-multipart>=0.0.9
